+++ dillon-module-infra/src/main/java/com/dillon/lw/module/infra/dal/dataobject/file/FileDO.java
@@ -1,16 +1,16 @@
 package com.dillon.lw.module.infra.dal.dataobject.file;
 
-import com.baomidou.mybatisplus.annotation.KeySequence;
-import com.baomidou.mybatisplus.annotation.TableName;
 import com.dillon.lw.framework.mybatis.core.dataobject.BaseDO;
 import com.dillon.lw.framework.tenant.core.aop.TenantIgnore;
+import com.baomidou.mybatisplus.annotation.KeySequence;
+import com.baomidou.mybatisplus.annotation.TableName;
 import lombok.*;
 
 /**
  * 文件表
  * 每次文件上传，都会记录一条记录到该表中
  *
- * @author liwen
+ * @author 芋道源码
  */
 @TableName("infra_file")
 @KeySequence("infra_file_seq") // 用于 Oracle、PostgreSQL、Kingbase、DB2、H2 数据库的主键自增。如果是 MySQL 等数据库，可不写。
@@ -29,7 +29,7 @@ public class FileDO extends BaseDO {
     private Long id;
     /**
      * 配置编号
-     * <p>
+     *
      * 关联 {@link FileConfigDO#getId()}
      */
     private Long configId;
@@ -52,6 +52,6 @@ public class FileDO extends BaseDO {
     /**
      * 文件大小
      */
-    private Integer size;
+    private Long size;
 
 }
